add superAdminEditRoleFilter

+++ src/main/java/com/guide_finder/servlet/role/DeleteRoleServlet.java
@@ -1,4 +1,5 @@
 package com.guide_finder.servlet.role;
+import com.guide_finder.model.user.Role;
 import com.guide_finder.service.abstraction.role.RoleService;
 import com.guide_finder.service.impl.RoleServiceImpl;
 import javax.servlet.ServletException;
@@ -7,6 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 
 
 //
@@ -20,8 +22,11 @@ public class DeleteRoleServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req,
                          HttpServletResponse resp) throws ServletException, IOException {
         int idRoleToDel = Integer.parseInt(req.getParameter("id"));
+        if (roleService.getRoleById(idRoleToDel) != null) {
+            roleService.deleteRole(idRoleToDel);
+        }
 
-        roleService.deleteRole(idRoleToDel);
+
         resp.sendRedirect("/admin/roles");
